allow CSV upload
