+++ applications/fishing-map/package.json
@@ -55,6 +55,7 @@
     "intersection-observer": "^0.11.0",
     "lodash": "^4.17.20",
     "luxon": "^1.25.0",
+    "papaparse": "^5.3.0",
     "qs": "^6.9.4",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
