+++ applications/fishing-map/src/features/datasets/NewDataset.tsx
@@ -1,274 +1,49 @@
 import React, { useState, useCallback, Fragment } from 'react'
-import cx from 'classnames'
-import { useDropzone } from 'react-dropzone'
+import type { FeatureCollectionWithFilename } from 'shpjs'
 import { useTranslation } from 'react-i18next'
 import lowerCase from 'lodash/lowerCase'
 import { useSelector } from 'react-redux'
-import type { FeatureCollectionWithFilename } from 'shpjs'
-import InputText from '@globalfishingwatch/ui-components/dist/input-text'
+import { parse as parseCSV } from 'papaparse'
 import Modal from '@globalfishingwatch/ui-components/dist/modal'
 import Button from '@globalfishingwatch/ui-components/dist/button'
-import Select from '@globalfishingwatch/ui-components/dist/select'
 import {
   AnyDatasetConfiguration,
   DatasetCategory,
   DatasetTypes,
   EnviromentalDatasetConfiguration,
 } from '@globalfishingwatch/api-types'
-import { ReactComponent as FilesIcon } from 'assets/icons/files-supported.svg'
+import { guessColumns } from '@globalfishingwatch/data-transforms'
 import { capitalize } from 'utils/shared'
 import { SUPPORT_EMAIL } from 'data/config'
 import { selectLocationType } from 'routes/routes.selectors'
 import { readBlobAs } from 'utils/files'
 import { useDatasetsAPI, useDatasetModalConnect, useNewDatasetConnect } from './datasets.hook'
 import styles from './NewDataset.module.css'
-
-const extractPropertiesFromGeojson = (geojson: GeoJSON.FeatureCollection) => {
-  if (!geojson?.features) return []
-  const uniqueProperties = Object.keys(
-    geojson.features.reduce(function (acc, { properties }) {
-      return { ...acc, ...properties }
-    }, {})
-  )
-  return uniqueProperties
-}
-
-const getPropertyRangeValuesFromGeojson = (
-  geojson: GeoJSON.FeatureCollection,
-  property: string
-) => {
-  if (!geojson?.features) return
-  const propertyRange = geojson.features.reduce(
-    function (acc, { properties }) {
-      const value = properties?.[property]
-      if (!value) return acc
-      return {
-        min: value < acc.min ? value : acc.min,
-        max: value > acc.max ? value : acc.max,
-      }
-    },
-    { min: 0, max: 0 }
-  )
-  return propertyRange
-}
-
-interface DatasetConfigProps {
-  className?: string
-  datasetCategory: DatasetCategory
-  onDatasetFieldChange: (field: DatasetMetadata | AnyDatasetConfiguration) => void
-  fileData: FeatureCollectionWithFilename
-  metadata: DatasetMetadata
-}
-
-const DatasetConfig: React.FC<DatasetConfigProps> = (props) => {
-  const { metadata, fileData, className = '', datasetCategory, onDatasetFieldChange } = props
-  const { t } = useTranslation()
-  const geojsonProperties = extractPropertiesFromGeojson(fileData)
-  const geojsonPropertiesOptions = geojsonProperties.map((property) => ({
-    id: property,
-    label: capitalize(property),
-  }))
-  const { min, max } =
-    (metadata.configuration as EnviromentalDatasetConfiguration)?.propertyToIncludeRange || {}
-  return (
-    <div className={cx(styles.datasetConfig, className)}>
-      <InputText
-        inputSize="small"
-        value={metadata.name}
-        label={t('common.name', 'Name')}
-        className={styles.input}
-        onChange={(e) => onDatasetFieldChange({ name: e.target.value })}
-      />
-      <InputText
-        inputSize="small"
-        label={t('common.description', 'Description')}
-        className={styles.input}
-        onChange={(e) => onDatasetFieldChange({ description: e.target.value })}
-      />
-      {/*
-      <Select
-        label={t('dataset.typeOfFeatures', 'Type of features')}
-        options={[]}
-        selectedOption={undefined}
-        onSelect={(selected) => {
-          console.log('selected', selected)
-        }}
-        onRemove={(removed) => {
-          console.log('removed', removed)
-        }}
-        direction="top"
-      />
-      <Select
-        label={t('dataset.timeField', 'Time field')}
-        options={[]}
-        selectedOption={undefined}
-        onSelect={(selected) => {
-          console.log('selected', selected)
-        }}
-        onRemove={(removed) => {
-          console.log('removed', removed)
-        }}
-        direction="top"
-      /> */}
-
-      {datasetCategory === DatasetCategory.Context && (
-        <Select
-          label={t('dataset.featuresNameField', 'Features name field')}
-          placeholder={t('selects.placeholder', 'Select an option')}
-          options={geojsonPropertiesOptions}
-          selectedOption={geojsonPropertiesOptions.find(
-            ({ id }) => id === metadata.configuration?.propertyToInclude
-          )}
-          onSelect={(selected) => {
-            onDatasetFieldChange({ propertyToInclude: selected.id })
-          }}
-          onRemove={() => {
-            onDatasetFieldChange({ propertyToInclude: undefined })
-          }}
-        />
-      )}
-      {datasetCategory === DatasetCategory.Environment && (
-        <div className={styles.row}>
-          <label className={styles.selectLabel}>
-            {t('dataset.colorByValue', 'Color features by value')}
-          </label>
-          <Select
-            className={styles.selectShort}
-            placeholder={t('selects.placeholder', 'Select an option')}
-            containerClassName={styles.selectContainer}
-            options={geojsonPropertiesOptions}
-            selectedOption={geojsonPropertiesOptions.find(
-              ({ id }) => id === metadata.configuration?.propertyToInclude
-            )}
-            onSelect={(selected) => {
-              // TODO: pre-populate min and max values depending on selection
-              onDatasetFieldChange({ propertyToInclude: selected.id })
-              const propertyRange = getPropertyRangeValuesFromGeojson(fileData, selected.id)
-              if (propertyRange) {
-                onDatasetFieldChange({
-                  propertyToIncludeRange: {
-                    min: propertyRange.min,
-                    max: propertyRange.max,
-                  },
-                })
-              }
-            }}
-            onRemove={() => {
-              onDatasetFieldChange({ propertyToInclude: undefined })
-            }}
-          />
-          <InputText
-            inputSize="small"
-            type="number"
-            step="0.1"
-            value={min}
-            placeholder={t('common.min', 'Min')}
-            className={cx(styles.shortInput, styles.noLabelInput)}
-            onChange={(e) =>
-              onDatasetFieldChange({
-                propertyToIncludeRange: {
-                  min: e.target.value && parseFloat(e.target.value),
-                  max:
-                    (metadata.configuration as EnviromentalDatasetConfiguration)
-                      ?.propertyToIncludeRange?.max || parseFloat(e.target.value),
-                },
-              })
-            }
-          />
-          <InputText
-            inputSize="small"
-            type="number"
-            step="0.1"
-            placeholder={t('common.max', 'Max')}
-            value={max}
-            className={cx(styles.shortInput, styles.noLabelInput)}
-            onChange={(e) =>
-              onDatasetFieldChange({
-                propertyToIncludeRange: {
-                  min: min || 0,
-                  max: e.target.value && parseFloat(e.target.value),
-                },
-              })
-            }
-          />
-        </div>
-      )}
-    </div>
-  )
-}
-
-interface DatasetFileProps {
-  className?: string
-  accept?: string
-  onFileLoaded: (fileInfo: File) => void
-}
-
-const DatasetFile: React.FC<DatasetFileProps> = ({ onFileLoaded, accept, className = '' }) => {
-  console.log(accept)
-  const { t } = useTranslation()
-  const onDropAccepted = useCallback(
-    (files) => {
-      onFileLoaded(files[0])
-    },
-    [onFileLoaded]
-  )
-  const { getRootProps, getInputProps, isDragActive, acceptedFiles, fileRejections } = useDropzone({
-    accept,
-    onDropAccepted,
-  })
-  console.log(acceptedFiles)
-  return (
-    <div className={cx(styles.dropFiles, className)} {...(getRootProps() as any)}>
-      <FilesIcon />
-      <input {...getInputProps()} />
-      {acceptedFiles.length ? (
-        <p className={styles.fileText}>
-          {t('dataset.file', 'File')}: {acceptedFiles[0].name}
-        </p>
-      ) : isDragActive ? (
-        <p className={styles.fileText}>{t('dataset.dragActive', 'Drop the file here ...')}</p>
-      ) : (
-        <p className={styles.fileText}>
-          {t('dataset.dragFilePlaceholder', {
-            defaultValue:
-              'Drag and drop a compressed shapefile or geojson here or click to select it',
-            interpolation: { escapeValue: false, useRawValueToEscape: true },
-          })}
-        </p>
-      )}
-      {fileRejections.length > 0 && (
-        <p className={cx(styles.fileText, styles.warning)}>
-          {t('dataset.fileNotAllowed', '(Only .zip or .json files are allowed)')}
-        </p>
-      )}
-    </div>
-  )
-}
+import DatasetFile from './DatasetFile'
+import DatasetConfig from './DatasetConfig'
 
 export type DatasetMetadata = {
   name: string
   description?: string
-  type: DatasetTypes.Context
+  type: DatasetTypes
   configuration?: AnyDatasetConfiguration
+  fields?: string[]
+  guessedFields?: Record<string, string>
 }
 
-type DatasetType = 'polygons' | 'tracks' | 'points'
+export type DatasetGeometryType = 'polygons' | 'tracks' | 'points'
 
-const ACCEPT_FILES_BY_TYPE: Record<DatasetType, string> = {
-  polygons: '.zip, .json, .geojson',
-  tracks: '.csv',
-  points: '',
-}
+export type CSV = Record<string, any>[]
 
 function NewDataset(): React.ReactElement {
   const { t } = useTranslation()
   const { datasetModal, datasetCategory, dispatchDatasetModal } = useDatasetModalConnect()
   const { addNewDatasetToWorkspace } = useNewDatasetConnect()
 
-  const [datasetType, setDatasetType] = useState<DatasetType | null>(null)
-  const [datasetTypeConfirmed, setDatasetTypeConfirmed] = useState<boolean>(false)
+  const [datasetGeometryType, setDatasetGeometryType] = useState<DatasetGeometryType | null>(null)
+  const [datasetGeometryTypeConfirmed, setDatasetGeometryTypeConfirmed] = useState<boolean>(false)
   const [file, setFile] = useState<File | undefined>()
-  const [fileData, setFileData] = useState<FeatureCollectionWithFilename | undefined>()
+  const [fileData, setFileData] = useState<FeatureCollectionWithFilename | CSV | undefined>()
   const [loading, setLoading] = useState(false)
   const [error, setError] = useState('')
   const [metadata, setMetadata] = useState<DatasetMetadata | undefined>()
@@ -276,60 +51,85 @@ function NewDataset(): React.ReactElement {
   const { dispatchCreateDataset } = useDatasetsAPI()
 
   const onFileLoaded = useCallback(
-    async (file: File) => {
+    async (file: File, type: DatasetGeometryType) => {
       setLoading(true)
       setError('')
-      const isZip =
-        file.type === 'application/zip' ||
-        file.type === 'application/x-zip-compressed' ||
-        file.type === 'application/octet-stream' ||
-        file.type === 'multipart/x-zip'
-      const isGeojson =
-        (!isZip && file.type === 'application/json') || (!isZip && file.name.includes('.geojson'))
       const name =
         file.name.lastIndexOf('.') > 0 ? file.name.substr(0, file.name.lastIndexOf('.')) : file.name
-      if (isGeojson && file.name.includes('.geojson')) {
-        const blob = file.slice(0, file.size, 'application/json')
-        const fileAsJson = new File([blob], `${name}.json`, { type: 'application/json' })
-        setFile(fileAsJson)
-      } else {
-        setFile(file)
-      }
 
-      let geojson: FeatureCollectionWithFilename | undefined = undefined
-      if (isZip) {
-        try {
-          const shpjs = await import('shpjs').then((module) => module.default)
-          const fileData = await readBlobAs(file, 'arrayBuffer')
-          // TODO support multiple files in shapefile
-          geojson = (await shpjs(fileData)) as FeatureCollectionWithFilename
-        } catch (e) {
-          console.warn('Error reading file:', e)
+      const metadataName = capitalize(lowerCase(name))
+
+      if (type === 'polygons') {
+        const isZip =
+          file.type === 'application/zip' ||
+          file.type === 'application/x-zip-compressed' ||
+          file.type === 'application/octet-stream' ||
+          file.type === 'multipart/x-zip'
+        const isGeojson =
+          (!isZip && file.type === 'application/json') || (!isZip && file.name.includes('.geojson'))
+
+        if (isGeojson && file.name.includes('.geojson')) {
+          const blob = file.slice(0, file.size, 'application/json')
+          const fileAsJson = new File([blob], `${name}.json`, { type: 'application/json' })
+          setFile(fileAsJson)
+        } else {
+          setFile(file)
         }
-      } else {
+
+        let geojson: FeatureCollectionWithFilename | undefined = undefined
+        if (isZip) {
+          try {
+            const shpjs = await import('shpjs').then((module) => module.default)
+            const fileData = await readBlobAs(file, 'arrayBuffer')
+            // TODO support multiple files in shapefile
+            geojson = (await shpjs(fileData)) as FeatureCollectionWithFilename
+          } catch (e) {
+            console.warn('Error reading file:', e)
+          }
+        } else {
+          const fileData = await readBlobAs(file, 'text')
+          try {
+            geojson = JSON.parse(fileData)
+          } catch (e) {
+            console.warn('Error reading file:', e)
+          }
+        }
+        if (geojson !== undefined) {
+          setFileData(geojson)
+          setMetadata((metadata) => ({
+            ...metadata,
+            name: metadataName,
+            type: DatasetTypes.Context,
+            category: datasetCategory,
+            configuration: {
+              // TODO when supporting multiple files upload
+              // ...(geojson?.fileName && { file: geojson.fileName }),
+              ...(isGeojson && { format: 'geojson' }),
+            },
+          }))
+        } else {
+          setFileData(undefined)
+          setError(t('errors.datasetNotValid', 'It seems to be something wrong with your file'))
+        }
+      } else if (type === 'tracks') {
+        setFile(file)
         const fileData = await readBlobAs(file, 'text')
-        try {
-          geojson = JSON.parse(fileData)
-        } catch (e) {
-          console.warn('Error reading file:', e)
-        }
-      }
-      if (geojson !== undefined) {
-        setFileData(geojson)
-        setMetadata((metadata) => ({
+        const { data, meta } = parseCSV(fileData, {
+          dynamicTyping: true,
+          header: true,
+          skipEmptyLines: true,
+        })
+        setFileData(data as CSV)
+        const guessedColumns = guessColumns(meta?.fields)
+        setMetadata({
           ...metadata,
-          name: capitalize(lowerCase(name)),
-          type: DatasetTypes.Context,
+          name: metadataName,
+          type: DatasetTypes.UserTracks,
           category: datasetCategory,
-          configuration: {
-            // TODO when supporting multiple files upload
-            // ...(geojson?.fileName && { file: geojson.fileName }),
-            ...(isGeojson && { format: 'geojson' }),
-          },
-        }))
-      } else {
-        setFileData(undefined)
-        setError(t('errors.datasetNotValid', 'It seems to be something wrong with your file'))
+          fields: meta!.fields,
+          guessedFields: guessedColumns,
+        } as any) // TODO
+        console.log(data, meta, guessedColumns)
       }
       setLoading(false)
     },
@@ -384,18 +184,17 @@ function NewDataset(): React.ReactElement {
     setLoading(false)
     setMetadata(undefined)
     dispatchDatasetModal(undefined)
+    setDatasetGeometryType(null)
+    setDatasetGeometryTypeConfirmed(false)
   }
 
   const onDatasetTypeChange = (e: any) => {
-    setDatasetType(e.target.value)
+    setDatasetGeometryType(e.target.value)
   }
 
   const onConfirmDatasetCategoryClick = () => {
-    setDatasetTypeConfirmed(true)
+    setDatasetGeometryTypeConfirmed(true)
   }
-
-  console.log(datasetType)
-  console.log(ACCEPT_FILES_BY_TYPE[datasetType as DatasetType])
 
   return (
     <Modal
@@ -409,7 +208,8 @@ function NewDataset(): React.ReactElement {
       onClose={onClose}
     >
       <div className={styles.modalContent}>
-        {datasetTypeConfirmed === false && datasetCategory === DatasetCategory.Environment ? (
+        {datasetGeometryTypeConfirmed === false &&
+        datasetCategory === DatasetCategory.Environment ? (
           <Fragment>
             <div onChange={onDatasetTypeChange}>
               <input type="radio" value="polygons" name="datasetType" /> Polygons
@@ -419,15 +219,13 @@ function NewDataset(): React.ReactElement {
           </Fragment>
         ) : (
           <Fragment>
-            <DatasetFile
-              onFileLoaded={onFileLoaded}
-              accept={ACCEPT_FILES_BY_TYPE[datasetType as DatasetType]}
-            />
+            <DatasetFile onFileLoaded={onFileLoaded} type={datasetGeometryType!} />
             {fileData && metadata && (
               <DatasetConfig
-                fileData={fileData}
+                fileData={fileData as FeatureCollectionWithFilename}
                 metadata={metadata}
                 datasetCategory={datasetCategory}
+                datasetGeometryType={datasetGeometryType!}
                 onDatasetFieldChange={onDatasetFieldChange}
               />
             )}
@@ -443,7 +241,7 @@ function NewDataset(): React.ReactElement {
             </a>
           </span>
         </div>
-        {datasetTypeConfirmed ? (
+        {datasetGeometryTypeConfirmed ? (
           <Button
             disabled={!file || !metadata?.name}
             className={styles.saveBtn}
@@ -454,7 +252,7 @@ function NewDataset(): React.ReactElement {
           </Button>
         ) : (
           <Button
-            disabled={!datasetType}
+            disabled={!datasetGeometryType}
             className={styles.saveBtn}
             onClick={onConfirmDatasetCategoryClick}
           >
