+++ packages/data-transforms/src/index.ts
@@ -1,2 +1,3 @@
 export * from './track-value-array-to-segments'
 export * from './segments'
+export * from './csv-to-track-segments'
